+++ Templates/src/PainKiller.PowerCommands.Bootstrap/whats_new.md
@@ -1,6 +1,10 @@
 ﻿# What is new?
+## Version 1.0.4.4
+**Released 2025-02-28**
+- Improved diagnostic logging
+- Minor fixes
 ## Version 1.0.4.3
-**Released 2025-01-27**
+**Released 2025-02-02**
 - New basic functionality for translation added.
 - Improved documentation functionality.
 - Added dialog for file and folder select to DialogService.